Fixing the unit test
