+++ lib/strategies/webapp-strategy.js
@@ -16,7 +16,7 @@ const Q = require("q");
 const acceptLanguage = require('accept-language');
 const constants = require('../utils/constants');
 const TokenUtil = require("../utils/token-util");
-const { request } = require('../utils/common-util');
+const request = require('../utils/request-util');
 const ServiceUtil = require('../utils/service-util');
 const appIdSupportedLanguagesArray = require("../data/languagesCodesArray");
