+++ test/webapp-strategy-test.js
@@ -27,7 +27,7 @@ describe("/lib/strategies/webapp-strategy", function () {
 	before(function () {
 		WebAppStrategy = proxyquire("../lib/strategies/webapp-strategy", {
 			"../utils/token-util": tokenUtilsMock,
-			"request": require("./mocks/request-mock")
+			"../utils/request-util": require("./mocks/request-mock")
 		});
 		webAppStrategy = new WebAppStrategy({
 			tenantId: "tenantId",
